Added testInt and testIntArray functions. The functions are implemented and work satisfactorily. testUInt and testUIntArray are next.

+++ src/ntDatabaseClient.cpp
@@ -38,7 +38,7 @@ using namespace epics::pvaClient;
 int main (int argc, char **argv)
 {
 	cout << "ntDatabase Client\n";
-	string types[] = {"string", "byte"};
+	string types[] = {"string", "int"};
 	int test_num = 2;
 	try {
